crisp up exported interfaces
TestFile is now file. TestSuite is now suite.

ContextAppendable is an interface introduced that is passed to the
Parser.Parse() function. Context is now exported and is the mechanism by
which context and state is passed around (no longer the TestFile
object). This results in a cleaner public interface and one which
doesn't require plugins to wrap gdt.TestFile.