add JSONPath format string validations
Adds the ability to specify a list of JSONPath expressions and the
associated format that the value at that expression should have.

For example, to validate that the "id" field of the JSON object in the
HTTP response is a valid UUID4 string, you can write:

```yaml
 - name: look up that created book
   GET: $LOCATION
   response:
     status: 200
     json:
       path_formats:
         $.id: uuid4
```

JSONPath format validation checks for the existence of the path and then
validates the format of the value in that path.

Currently the only validated format strings are "uuid" and "uuid4".
Passing a different format string in the test file results in a failure,
like so:

```
jaypipes@thelio:~/go/src/github.com/jaypipes/gdt/examples/books/tests/api$ go test -v
=== RUN   TestBooksAPI
=== RUN   TestBooksAPI/create_then_get.yaml
=== RUN   TestBooksAPI/create_then_get.yaml/create_a_new_book
=== RUN   TestBooksAPI/create_then_get.yaml/look_up_that_created_book
=== RUN   TestBooksAPI/failures.yaml
=== RUN   TestBooksAPI/failures.yaml/no_such_book_was_found
=== RUN   TestBooksAPI/failures.yaml/invalid_query_parameter_is_supplied
--- FAIL: TestBooksAPI (0.00s)
    --- FAIL: TestBooksAPI/create_then_get.yaml (0.00s)
        --- PASS: TestBooksAPI/create_then_get.yaml/create_a_new_book (0.00s)
        --- FAIL: TestBooksAPI/create_then_get.yaml/look_up_that_created_book (0.00s)
            file.go:111:
                	Error Trace:	assertions.go:115
                	            				assertions.go:104
                	            				assertions.go:70
                	            				file.go:111
                	Error:      	Expected nil, but got: &errors.errorString{s:"unknown format uui"}
                	Test:       	TestBooksAPI/create_then_get.yaml/look_up_that_created_book
                	Messages:   	Unknown format uui in test
    --- PASS: TestBooksAPI/failures.yaml (0.00s)
        --- PASS: TestBooksAPI/failures.yaml/no_such_book_was_found (0.00s)
        --- PASS: TestBooksAPI/failures.yaml/invalid_query_parameter_is_supplied (0.00s)
FAIL
exit status 1
FAIL	github.com/jaypipes/gdt/examples/books/tests/api	0.007s
```