reduce os.Open() calls during parse operations
converts the exported `gdt.Parse` function into an unexported
`gdt.file:File.parse` method that accepts an io.Reader instead of a
filepath string. This means we can reduce duplicative os.Open calls
since the `gdt.From` function already os.Open'd the test file to
determine if it's a directory or not.