simplify gdt.http file parsing
refactors the way that test suites and test cases are parsed and
constructed, ensuring that testing.T.Run() is called with a valid string
name and not "#00" or similar. Adds `NewTestSuiteFromDir`,
`NewTestCaseFromReader`, `NewTestCaseFromBytes` functions that return
TestSuite and TestCase structs, respectively, leaving the `From`
function as the primary entry point for most users.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>